requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --allow-unsafe --output-file=requirements-dev.txt requirements-dev.in
#
-e git+https://github.com/Rapptz/discord.py@master#egg=discord.py
    # via
    #   -r requirements.in
    #   jishaku
    #   kb1000-discordpy-common
aiohttp==3.7.4.post0
    # via discord.py
asgiref==3.3.1
    # via django
astroid==2.5.2
    # via pylint
astunparse==1.6.3
    # via import-expression
async-timeout==3.0.1
    # via aiohttp
asyncpg==0.22.0
    # via -r requirements.in
attrs==20.3.0
    # via aiohttp
braceexpand==0.1.6
    # via jishaku
chardet==4.0.0
    # via aiohttp
click==7.1.2
    # via pip-tools
django==3.1.7
    # via -r requirements.in
humanize==3.3.0
    # via jishaku
idna==3.1
    # via yarl
import-expression==1.1.4
    # via jishaku
isort==5.8.0
    # via pylint
jedi==0.18.0
    # via -r requirements-dev.in
jishaku==1.20.0.220
    # via -r requirements.in
kb1000-discordpy-common==1.3
    # via -r requirements.in
lazy-object-proxy==1.6.0
    # via astroid
mccabe==0.6.1
    # via pylint
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
parso==0.8.2
    # via jedi
pep517==0.10.0
    # via pip-tools
pip-tools==6.0.1
    # via -r requirements-dev.in
psutil==5.8.0
    # via -r requirements.in
psycopg2==2.8.6
    # via -r requirements.in
pylint==2.7.4
    # via -r requirements-dev.in
pytz==2021.1
    # via django
pyyaml==5.4.1
    # via -r requirements.in
six==1.15.0
    # via astunparse
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via
    #   pep517
    #   pylint
typing-extensions==3.7.4.3
    # via aiohttp
wheel==0.36.2
    # via
    #   -r requirements-dev.in
    #   astunparse
wrapt==1.12.1
    # via astroid
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
pip==21.0.1
    # via
    #   -r requirements-dev.in
    #   pip-tools
setuptools==54.2.0
    # via
    #   -r requirements-dev.in
    #   humanize
